Refactor availability component into template-specific variants
